refactor: extract agent command handlers
